Moved GuiElectricityMeter to new machine package.
