+++ src/main/java/org/halvors/electrometrics/common/base/MachineType.java
@@ -4,7 +4,7 @@ import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import org.halvors.electrometrics.Electrometrics;
-import org.halvors.electrometrics.client.gui.GuiElectricityMeter;
+import org.halvors.electrometrics.client.gui.machine.GuiElectricityMeter;
 import org.halvors.electrometrics.common.tile.TileEntityMachine;
 import org.halvors.electrometrics.common.tile.machine.TileEntityElectricityMeter;
 import org.halvors.electrometrics.common.util.LanguageUtils;
